% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_fishing.R
\name{calcBiomass}
\alias{calcBiomass}
\title{Total Biomass Calculation}
\usage{
calcBiomass(sim, etaTime = 0.4)
}
\arguments{
\item{sim}{The FEISTY simulation result list from \code{\link{simulateFEISTY}}.}

\item{etaTime}{The last fraction of the simulation period (default the last 40\%).}
}
\value{
Add SSB data to the result list:
\itemize{
\item BiomassMean: a vector containing the arithmetic mean SSB data [g/m2/year] of each functional type of the time range specified.
\item BiomassGMean: a vector containing the natural log-based geometric mean SSB data [g/m2/year] of each functional type of the time range specified.
\item BiomassMin: a vector containing the minimum SSB data [g/m2/year] of each functional type within the time range specified.
\item BiomassMax: a vector containing the maximum SSB data [g/m2/year] of each functional type within the time range specified.
\item Biomass: a matrix containing the SSB data [g/m2/year] of each functional type (column) in each time point (row)
}
}
\description{
This function calculates the total (B) for each functional 
type based on a FEISTY simulation result in units of g/m2. 
The function averages over a certain period, determined by etaTime
}
\details{
This function calculates the biomass for each function type based on the FEISTY simulation results within the specified time fraction (default is the last 40\% of the simulation period). \cr
}
\seealso{
\code{\link{paramAddGroup}} 	Add parameters of one functional type
}
\author{
DDDD
}
