% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_fishing.R
\name{calcYield}
\alias{calcYield}
\title{Yield Calculation}
\usage{
calcYield(sim, etaTime = 0.4)
}
\arguments{
\item{sim}{The FEISTY simulation result list from \code{\link{simulateFEISTY}}.}

\item{etaTime}{The last fraction of the simulation period (default the last 40\%).}
}
\value{
Add yield data to the result list:
\itemize{
\item yieldAMean: a vector containing the arithmetic mean yield data [g/m2/year] of each functional type of the time range specified.
\item yieldGMean: a vector containing the natural log-based geometric mean yield data [g/m2/year] of each functional type of the time range specified.
\item yieldMin: a vector containing the minimum yield data [g/m2/year] of each functional type within the time range specified.
\item yieldMax: a vector containing the maximum yield data [g/m2/year] of each functional type within the time range specified.
\item yield: a matrix containing the yield data [g/m2/year] of each functional type (column) in each time point (row)
}
}
\description{
This function calculates the yield for each functional type in units of g/m2/year based on a FEISTY simulation result.
}
\details{
This function calculates the yield for each function type based on the FEISTY simulation results within the specified time fraction (default is the last 40\% of the simulation period). \cr
Yield is the product of biomass \code{u} [g/m2] and fishing mortality \code{mortF} [1/year]. Negative values in \code{u} are corrected to 0. 
The yield calculation is automatically adaptive to non-time-series simulations or time-series simulations. \cr
This function is called automatically inside \code{\link{simulateFEISTY}} (with \code{etaTime = 0.4}),
so its outputs are already present in a standard simulation result. It can also be called directly on a
simulation result, for example to recompute the yield over a different time fraction \code{etaTime}.
}
\seealso{
\code{\link{setFishing}} 	Set fishing mortality
}
\author{
Yixin Zhao
}
