% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_main.R
\name{derivativesFEISTYR}
\alias{derivativesFEISTYR}
\title{Calculate Derivatives of State Variables in FEISTY Model in R}
\usage{
derivativesFEISTYR(t, u, p, FullOutput=TRUE)
}
\arguments{
\item{t}{Current time. Any numeric value, no use currently.}

\item{u}{A numeric vector containing all state variables, all resources + all size classes.}

\item{p}{The parameter list.}

\item{FullOutput}{Logical flag indicating whether to return all calculation results. Default TRUE.}
}
\value{
If \code{FullOutput} is 'TRUE', a list containing the following components:
\itemize{
\item deriv: a vector of derivatives [g/m2/year] of all resources and all functional type size classes.
\item f: a vector containing feeding levels [-] of all resources (0) and all size classes of functional types.
\item glvl: a vector containing oxygen limitation levels [-] of all size classes of functional types. Resources not included.
\item mortpred: a vector containing predation mortality rate [1/year] of all resources and all size classes of functional types.
\item g: Net growth rate (the fraction of available energy invested in growth) [1/year]. It includes all size classes of functional types. Resources not included.
\item Repro: Energy used for reproduction of all size classes of functional types [g/m2/year]. Resources not included.
\item Fin: Biomass flux into each size class [g/m2/year]. Resources not included.
\item Fout: Biomass flux out of each size class [g/m2/year]. Resources not included.
\item totMort: a vector containing total mortality [g/m2/year] of each functional type, 
which includes predation mortality, background mortality, and fishing mortality.
\item totGrazing: a vector containing total grazing (food intake before assimilation) [g/m2/year] of each functional type. glvl * Cmax * f * u (oxygen limitation * maximum consumption rate * feeding level * biomass).
\item totLoss: a vector containing total biomass loss [g/m2/year] of each functional type, including unassimilated food, basal metabolism and reproduction cost (1-epsRepro). 
The reproduction cost here also include the energy loss (1-epsRepro) from the flux out of the last size class (invested in reproduction) of each functional type.
These losses are supposed to be released to environments.
\item totRepro: a vector containing total energy used for reproduction [g/m2/year] of each functional type (before multiplying the reproduction efficiency epsRepro).
\item totRecruit: a vector containing total recruitment [g/m2/year] of each functional type. 
totRecruit = totRepro * epsRepro (reproduction efficiency)
\item totBiomass: a vector containing total biomass [g/m2] of each functional type. 
}
If \code{FullOutput = FALSE}, it only returns derivatives of resources and all size classes.
}
\description{
This function calculates the derivatives of state variables in FEISTY model in R. 
It is used for time integration of FEISTY simulations or getting rates by computing derivatives of one time step.
Note that calculating derivatives in R is much slower than with Fortran.
}
\details{
This function is designed to simulate the dynamics of marine resources and fish populations in FEISTY model in R. 
The derivative calculation of resources can be based on different types of dynamics (chemostat or logistic).
Negative state variables and NA values are revised to zero to prevent computational errors.
ODE solving is conducted by package deSolve. The equations are documented in the vignette.
More information can be found in de Roos et al. (2008) and Petrik et al. (2019).
}
\examples{
# Parameter list preparation
p = setupBasic2()
# FEISTY simulation 200 years, use derivativesFEISTYR
sim = simulateFEISTY(p=p, tEnd=200, USEdll=FALSE)
# get rates of last year, based on biomass of last year 
rates = derivativesFEISTYR(t=0, p=p, u=sim$u[sim$nTime,], FullOutput=TRUE)

}
\references{
Petrik, C. M., Stock, C. A., Andersen, K. H., van Denderen, P. D., & Watson, J. R. (2019). Bottom-up drivers of global patterns of demersal, forage, and pelagic fishes. Progress in oceanography, 176, 102124.

van Denderen, P. D., Petrik, C. M., Stock, C. A., & Andersen, K. H. (2021). Emergent global biogeography of marine fish food webs. Global Ecology and Biogeography, 30(9), 1822-1834.

de Roos, A. M., Schellekens, T., Van Kooten, T., Van De Wolfshaar, K., Claessen, D., & Persson, L. (2008). Simplifying a physiologically structured population model to a stage-structured biomass model. Theoretical population biology, 73(1), 47-62.
}
\seealso{
\code{\link{simulateFEISTY}} The main function for running FEISTY simulations
}
\author{
Ken H. Andersen, Karline Soetaert, Yixin Zhao
}
