\name{FeistyR-package}
\alias{FeistyR-package}
\alias{FeistyR}
\docType{package}
\title{
  FEISTY: FishErIes Size and functional TYpe model. 
}
\description{
The FEISTY model is a size- and trait-based fish model, described in Petrik et al., 2019 and van Denderen et al., 2020.
     }
\details{
  \tabular{ll}{
    Package: \tab FeistyR\cr
    Type: \tab Package\cr
    Created: \tab 2022-31-01\cr
    License: \tab GNU Public License 2 or above\cr
  }
}
\author{
  Ken H Andersen,
  Karline Soetaert, 
  Yixin Zhao
}
\seealso{
\itemize{
 \item \code{\link{simulateFeisty}} Main function for simulations of the FEISTY model.
 %\item \code{\link{setupPelagicSpecies}} Parameter input of the feisty model
}
}
\examples{
\dontrun{
## show examples (see respective help pages for details)
example(simulateFeisty)
}
}

\references{
Petrik, C. M., Stock, C. A., Andersen, K. H., van Denderen, P. D., & Watson, J. R. (2019). Bottom-up drivers of global patterns of demersal, forage, and pelagic fishes. Progress in oceanography, 176, 102124.

van Denderen, P. D., Petrik, C. M., Stock, C. A., & Andersen, K. H. (2021). Emergent global biogeography of marine fish food webs. Global Ecology and Biogeography, 30(9), 1822-1834.

}
\keyword{ package }
