% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_parms.R
\name{paramAddGroup}
\alias{paramAddGroup}
\title{Add parameters of a functional type}
\usage{
paramAddGroup(p, nStages, mMin, mMax, mMature=NA, mortF=0, mort0=0.1, u0=1, name=NA)
}
\arguments{
\item{p}{The parameter list to be updated.}

\item{nStages}{Number of stages for this functional type.}

\item{mMin}{Minimum size (mass) of this functional type (boundary value), in gram wet weight [gWW].}

\item{mMax}{Maximum size (mass) of this functional type (boundary value), in gram wet weight [gWW].}

\item{mMature}{Size [gWW] at which fish has a 50\% maturity level, which will be used for maturity level calculation based on an S-shape function.\cr
If NA, only the last size class is 50\% mature; others are 0.}

\item{mortF}{Fishing mortality for all size classes [1/year]. Default value 0, indicating no fishing.}

\item{mort0}{Natural mortality (background mortality) for all size groups [1/year]. Default value 0.1.}

\item{u0}{Initial biomass value of all size classes [gWW/m2]. Default value 1.}

\item{name}{The name (acronym) of the functional type. If not provided, a default name is assigned.}
}
\value{
The updated parameter list \code{p}:
\itemize{
\item nGroups, number of all functional types
\item groupnames, a character vector of names of resources and each fish functional groups
\item stagenames, a character vector of names of resources and each fish stage (functional groups)
\item ix, a list including sublists which contains indices of the size classes of each functional type.
\item mLower, a vector containing the lower limit of size of each size class of resource and functional type
\item mUpper, a vector containing the upper limit of size of each size class of resource and functional type
\item z, a vector containing the ratio between \code{mUpper} and \code{mLower}, all resources (NA) + all size classes
\item mc, a vector containing the geometric mean size of resources and all stages of functional types
\item mMature, a vector containing the size with a 50\% maturity level of each functional type, from parameter input
\item psiMature, a vector containing the maturity level, all resources (NA) + all size classes
\item mortF, a vector containing fishing mortality, all resources (NA) + all size classes
\item mort0, a vector containing all background mortality, all resources (NA) + all size classes
\item u0, a vector containing the initial value of biomass, all resources (from \code{\link{paramAddResource}}) + all size classes
\item ixFish, a vector of indices of all functional type size classes, starting from \code{nResources+1}
\item nStages, an integer, total number of resources and size classes.
}
All the parameters above will be updated if a new functional type is added by another call of \code{paramAddGroup}. \cr
NAs in \code{z}, \code{psiMature}, \code{mortF}, and \code{mort0} will be revised to 0 in \code{\link{paramAddPhysiology}}.
}
\description{
This function updates the parameter list by adding parameters of one functional type.
}
\details{
This function is designed to add parameters of a functional type to the parameter list.
Generally, this function needs to be called after the function \code{\link{paramAddResource}}.
Every call of this function can add \bold{one} functional group, which means 
this function needs to be called multiple times to add all functional types.
}
\examples{
# Just an example, data may not make sense.
p <- paramInit()
# add four resources
p <- paramAddResource(p,
     names= c("smallZoo", "largeZoo", "smallBenthos", "largeBenthos"),
     K    = c(100, 120, 80, 0),
     r    = c(1, 1, 1, 1),
     mLower = c(2e-06,0.001, 0.5e-03, 0.25),
     mUpper = c(0.001, 0.5, 125, 125),
     mc   = c(2e-06*sqrt(500), 0.001*sqrt(500), 0.5e-03*sqrt(250000), 0.25*sqrt(500)))
# add the first functional type
p <- paramAddGroup(p, nStages = 3, mMin = 0.1, mMax = 100, mMature = 100*0.25, mortF=0, mort0 = 0.1, name = "smallfish")
# add the second functional type
p <- paramAddGroup(p, nStages = 6, mMin = 0.1, mMax = 100000, mMature = 100000*0.25, mortF=0, mort0 = 0.1, name = "largefish")

}
\seealso{
\code{\link{paramAddPhysiology}} Add Physiological Parameters

\code{\link{paramSizepref}} Size Preference Matrix Calculation
}
\author{
Ken H. Andersen, Karline Soetaert <karline.soetaert@nioz.nl>, Yixin Zhao
}
