% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_parms.R
\name{paramInit}
\alias{paramInit}
\title{Initialize Parameters for FEISTY}
\usage{
paramInit(...)
}
\arguments{
\item{...}{Additional parameters needed to be added to the list, which will be included in the returned parameter list.}
}
\value{
A parameter list is initialized with some basic blank values, vectors, matrices, and sublists.
\itemize{
\item ..., as input parameters in \code{paramInit(...)}
\item nResources, the total number of resources, will be updated later
\item nGroups, the total number of fish functional groups, will be updated later
\item nStages, the total number of resources + fish stages, will be updated later
\item groupnames, a character vector of names of resources and each fish functional group, will be updated later
\item stagenames, a character vector of names of resources and each fish stage (functional group), will be updated later
\item theta, a matrix for the feeding preference of each predator x to each prey y, will be updated later
\item ix, index lists for size classes of each fish functional type (sublists), will be updated later
\item ixFish, a vector containing indices for all fish, will be updated later
\item ixR, a vector containing indices for all resources, will be updated later
\item my_palette, default color palettes for all functional types
\item my_names, default full name and acronyms of all functional types
}
}
\description{
This function initializes a list of parameters for the FEISTY model.
}
\details{
This function prepares some basic values, vectors, matrices, and sublists for the following parameter setting up.
Default color palettes and names of all resources and functional types (all stages) are defined, which will be used in visualization.
The returned parameter list will be used for further updates.
}
\examples{
# Initialize parameters
p <- paramInit(size_1 = 500, size_2 = 1000, depth_1 = 200, depth_2 = 250) # add four extra parameters to the list

}
\seealso{
\code{\link{paramAddResource}} Add Resource Parameters
}
\author{
Ken H. Andersen, Karline Soetaert <karline.soetaert@nioz.nl>, Yixin Zhao
}
