% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_parms.R
\name{paramSizepref}
\alias{paramSizepref}
\title{Size Preference Matrix Calculation}
\usage{
paramSizepref(p, beta = 400, sigma = 1.3, type = 1)
}
\arguments{
\item{p}{The parameter list to be updated.}

\item{beta}{The preferred predator/prey mass ratio. Default 400.}

\item{sigma}{The width of size preference for feeding. Default 1.3.}

\item{type}{The type of size preference function. 1 for a normal distribution 
based function, 2 for a function based on error function, and 3 for a 
more complex function integrating over both predator and prey size groups.
Default 1.}
}
\value{
theta: feeding preference matrix for each predator x to each prey y.
}
\description{
This function calculates a size preference matrix of a predator for a prey based on the size.
}
\details{
This function computes a size preference matrix \code{theta}.
Each value represents the preference of a predator (row) for a prey (column) based on their sizes. \cr
Three types of functions can be selected:
\itemize{
\item Type 1 (normal): Based on a normal distribution around the preferred size ratio.
\item Type 2 (error function): Based on the error function as described in van Denderen et al., 2020.
\item Type 3 (double integration): Based on Andersen and Visser 2023,
this method integrates over both predator and prey size groups for a detailed preference calculation.
}

The preference is set to 0 for prey larger than the predator.

This function only needs to be called once in each simulation. It needs to be called after all functional types are added, i.e., after the last call of the function \code{\link{paramAddGroup}}.
The returned parameter list can be used for further updates.
}
\examples{
# Just an example, data may not make sense.
p <- paramInit()
# add three resources
p <- paramAddResource(p,
     names= c("smallZoo", "largeZoo", "smallBenthos"),
     K    = c(100, 120, 80),
     r    = c(1, 1, 1),
     mLower = c(2e-06,0.001, 0.5e-03),
     mUpper = c(0.001, 0.5, 125),
     mc   = c(2e-06*sqrt(500), 0.001*sqrt(500), 0.5e-03*sqrt(250000)))
# add the first functional type
p <- paramAddGroup(p, nStages = 3, mMin = 0.1, mMax = 100, mMature = 100*0.25, mortF=0, mort0 = 0.1, name = "smallfish")
# add the second functional type
p <- paramAddGroup(p, nStages = 6, mMin = 0.1, mMax = 100000, mMature = 100000*0.25, mortF=0, mort0 = 0.1, name = "largefish")
# add the third functional type
p <- paramAddGroup(p, nStages = 9, mMin = 0.1, mMax = 500000, mMature = 500000*0.25, mortF=0, mort0 = 0.1, name = "giantfish")
# add physiological parameters for three functional types (smallfish, largefish and giantfish)
p <- paramAddPhysiology(p, 
     ac = 20, bc = -0.25,       
     am = 0.011*365, bm = -0.175,      
     ae = 70, be = -0.2,        
     epsRepro = 0.01, 
     epsAssim = 0.7)
# Calculate size preference matrix
p$theta <- paramSizepref(p=p,
           beta = 400,
           sigma = 1.3,
           type = 3)
# Run the simulation for this customized FEISTY setup.           
 sim=simulateFEISTY(bCust=TRUE,p=p, tEnd=100)
# Plot dynamics of resources and small fish.
 plot(sim, sim, which=1:6, lty=1, mfrow=c(2,3), subset=time >= 0)       

}
\references{
van Denderen, P. D., Petrik, C. M., Stock, C. A., & Andersen, K. H. (2021). Emergent global biogeography of marine fish food webs. Global Ecology and Biogeography, 30(9), 1822-1834.

Andersen, K. H., & Visser, A. W. (2023). From cell size and first principles to structure and function of unicellular plankton communities. Progress in Oceanography, 213, 102995.
}
\author{
Yixin Zhao
}
