% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_parms.R
\name{paramTeffect}
\alias{paramTeffect}
\title{Add temperature effects}
\usage{
paramTeffect(p, Tref, Q10, Q10m, pelgroupidx, demgroupidx)
}
\arguments{
\item{p}{A parameter list. Must be used after \code{\link{paramAddPhysiology}}. \cr
`Tp` and `Tb` must be included. See what are `Tp` and `Tb` in \code{\link{setupBasic}} and \code{\link{setupBasic2}}.}

\item{Tref}{Reference temperature. Default 10 Celsius, generally cannot be other values, unless users define physiological rates based on another reference temperature.}

\item{Q10}{Q10 factor for the maximum consumption rate \code{Cmax} and clearance rate \code{V} [-]}

\item{Q10m}{Q10 factor for metabolism rates \code{metabolism} [-]}

\item{pelgroupidx}{pelagic fish group indices. If there is no pelagic fish, please assign NA.}

\item{demgroupidx}{demersal fish group indices. If there is no demersal fish, please assign NA.}
}
\value{
An updated parameter list:
\itemize{
\item Tref, from parameter input
\item Q10, from parameter input
\item Q10m, from parameter input
\item pelgroupidx, pelagic fish group indices, from parameter input
\item demgroupidx, demersal fish group indices, from parameter input
\item fT: factor of T effects on \code{V} and \code{Cmax} of pelagic fish.
\item fT_met: factor of T effects on \code{metabolism}.
\item fT_dem: factor of T effects on \code{V} and \code{Cmax} of all demersal fish in deep water (depth >= 200m), and small and medium demersal fish in shallow water (depth < 200m).
\item fT_met_dem: factor of T effects on \code{metabolism} of all demersal fish in deep water (depth >= 200m), and small and medium demersal fish in shallow water (depth < 200m).
\item fT_dem_shallow: factor of T effects on \code{V} and \code{Cmax} of large demersal fish in shallow water (depth < 200m).
\item fT_met_dem_shallow: factor of T effects on the \code{metabolism} of large demersal fish in shallow water (depth < 200m).
\item smdemidx: small demersal fish grid indices.
\item lgdemidx: large demersal fish grid indices.
\item pelgrididx: all pelagic prey grid indices, including zooplankton resources, pelagic fish (based on `pelgroupidx`) and small demersal fish (`smdemidx`).
\item allgrididx: all state variable grids, 1:nGrid.
}
}
\description{
This function adjusts temperature-dependent physiological rates based on environmental temperatures.\cr
It also prepares some indices for effective temperature effects on large demersal fish in demersal water.\cr
It applies Q10 temperature coefficients to modify metabolic rates, clearance rates, and maximum consumption rates.
}
\details{
The function calculates temperature effects on physiological rates (metabolism, clearance rate, and maximum consumption rate) using 
the Q10 theory, which states that a physiological process rate increases by a factor of Q10 for every 10 Celsius rise in temperature.
The effects are applied differently in terms of living habitats (pelagic and demersal) and water depth (shallow and deep).

The function requires a specific structure in the input parameter list `p`, such as
indices of all fish and vectors for saved baseline values of physiological rates (\code{Vsave}, \code{Cmaxsave}, \code{metabolismsave}).
Therefore it only can be called after \code{\link{paramAddPhysiology}}.
`Tp` and `Tb` are required in the input parameter list, or this function will not work.
It might not work in customized setups other than \code{\link{setupBasic}} and \code{\link{setupBasic2}}.
}
\examples{
# Just an example, data may not make sense.
# Create a FEISTY setup by setupBasic2.
p1 = setupBasic2(szprod = 100, lzprod = 100, bprodin = 5, depth = 100, Tp = 10, Tb = 8, 
nStages=9, etaMature=0.25, F=0, etaF=0.05)

# change environmental temperatures and update T effects on physiological rates
p1$Tp=15
p1$Tb=12
p2=paramTeffect(p1, p1$Tref, p1$Q10, p1$Q10m, pelgroupidx=c(1,2), demgroupidx=3)
# Compare the original and updated physiological rates 
p1$V
p2$V
p1$metabolism
p2$metabolism
p1$Cmax
p2$Cmax

}
\references{
Petrik, C. M., Stock, C. A., Andersen, K. H., van Denderen, P. D., & Watson, J. R. (2019). Bottom-up drivers of global patterns of demersal, forage, and pelagic fishes. Progress in oceanography, 176, 102124.

van Denderen, P. D., Petrik, C. M., Stock, C. A., & Andersen, K. H. (2021). Emergent global biogeography of marine fish food webs. Global Ecology and Biogeography, 30(9), 1822-1834.
}
\seealso{
\code{\link{setupBasic}} The setup following Petrik et al. (2019) \cr
\code{\link{setupBasic2}} A revised setup based on `setupBasic`
}
\author{
Yixin Zhao
}
