% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plot.FEISTY}
\alias{plot.FEISTY}
\title{Plot FEISTY simulation results (S3)}
\usage{
\method{plot}{FEISTY}(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
This function is a shortcut for \code{\link{plotSimulation}}.
}
\details{
See \code{\link{plotSimulation}}.
}
\examples{
sim=simulateFEISTY()
plot(sim)

}
