% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotDiet}
\alias{plotDiet}
\title{Diet plot}
\usage{
plotDiet(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Revised Based on the work of Daniel Ottmann Riera and Solenne Roux.
This function only works on four prepared setups or revised versions based on these four setups.
If customized setups by users have massive differences e.g., more resources and fish functional types, this plot function may not work.
Users may need to update scripts based on source codes.
}
\examples{
sim=simulateFEISTY()
plotDiet(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
Yixin Zhao
}
