% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feisty_plotting.R
\name{plotRates}
\alias{plotRates}
\title{Plots for growth rate, mortality, and feeding level}
\usage{
plotRates(p=sim$p, y=p$u0, bPlot=TRUE)
}
\arguments{
\item{p}{The parameter set \code{sim$p}. The \code{simpleOutput} must be TRUE in simulateFeisty().}

\item{y}{The state variable vector. For example, \code{sim$u[,sim$nTime]}. This means the biomass data in the last time point is used for rate calculation.
The time-averaged biomass data might be preferred, \code{colMeans(sim$u[,round(0.6*iTime):sim$nTime])}.}

\item{bPlot}{Boolean option determining whether to create a new plot panel. \cr 
If TRUE, create a new independent plot. \cr 
If FALSE, users need to define the subplot layout first by \code{defaultplot(mfcol=c(x,y))}. See an example \code{\link{plotSimulation}} in Feisty_plotting.R.}
}
\description{
Plot growth rate (1/year), mortality (1/year), and feeding level (dimensionless) over the size spectrum.
Data is calculated based on averaged biomass over the last 40% simulation time.
}
\examples{
sim=simulateFeisty(simpleOutput=TRUE)
plotRates(p=sim$p, y=sim$u[,sim$nTime], bPlot=TRUE)

}
\seealso{
\code{\link{simulateFeisty}} Run FEISTY model simulations \cr
\code{\link{derivativesFeistyR}} The derivative function of FEISTY
}
\author{
Yixin Zhao
}
