% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feisty_plotting.R
\name{plotSimulation}
\alias{plotSimulation}
\title{Plot simulation results}
\usage{
plotSimulation(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results. The \code{simpleOutput} must be TRUE in simulateFeisty().}
}
\description{
Make a plot combo for the simulation results, including \code{\link{plotSSBtime}}, \code{\link{plotSpectra}}, and \code{\link{plotRates}}.\cr
}
\details{
This function is designed for \code{\link{webFeisty}}.\cr 
There is a function \code{addLegends} in Feisty_plotting.R to add legends.\cr
Users may need to define their new scripts for adding legends according to the function \code{\link{addLegends}}.
}
\examples{
sim=simulateFeisty(simpleOutput=TRUE)
plotSimulation(sim)

}
\seealso{
\code{\link{simulateFeisty}} Run FEISTY model simulations \cr
\code{\link{plotSSBtime}} Spawning stock biomass plot \cr
\code{\link{plotSpectra}} Biomass plot \cr
\code{\link{plotRates}} Plots for growth rate, mortality, and feeding level
}
\author{
Yixin Zhao
}
