% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Biomass spectra plot}
\usage{
plotSpectra(sim, norm = T)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}

\item{norm}{Logic flag, controlling whether to plot the normalized biomass. The default is TRUE.}
}
\description{
Makes a plot of the biomass of all functional types over the size spectrum.
Data is averaged over the last 40\% simulation time.
}
\details{
The X-axis is individual weight on a log10 scale. The Y-axis is 
biomass on a log10 scale. If the logical flag "norm" is set to TRUE, the biomass is normalized by dividing it by \eqn{log(m_{u,i}/m_{l,i}}, 
where \eqn{m_{u,i}} and \eqn{m_{l,i}} are upper and lower boundary sizes of the size bin \eqn{i}, respectively.
See Box V in Andersen and Visser (2023).
}
\examples{
sim=simulateFEISTY()
plotSpectra(sim, norm=TRUE)
plotSpectra(sim, norm=FALSE)

}
\references{
Andersen, K. H., & Visser, A. W. (2023). From cell size and first principles to structure and function of unicellular plankton communities. Progress in Oceanography, 213, 102995.
}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
P. Daniël van Denderen, Yixin Zhao
}
