% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{PlotSpectraGgplot}
\alias{PlotSpectraGgplot}
\title{Plot Spectra with Ggplot}
\usage{
PlotSpectraGgplot(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Create spectrum plot used in \code{\link{plotSimulationPaper}}.
}
\details{
This function produces 3 pannels a. biomass, b. mortality and c. feeding level simulated by FEISTY for each group and size
}
\examples{
sim=simulateFEISTY()
PlotSpectraGgplot(sim)

}
\seealso{
\code{\link{webFEISTY}} A shiny interface for visualizing FEISTY model results \cr
\code{\link{simulateFEISTY}} Run FEISTY model simulations \cr
\code{\link{plotSimulationPaper}} Plot Fig.2 from publication  \cr
\code{\link{plotTimeSeries}} Plot Time series \cr
\code{\link{plotNetwork}} Food web plot
}
\author{
who ever think they should be authors R.D.
}
