% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feisty_plotting.R
\name{plotSSBtime}
\alias{plotSSBtime}
\title{Spawning stock biomass plot}
\usage{
plotSSBtime(sim, bPlot=TRUE)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results. The \code{simpleOutput} must be TRUE in simulateFeisty().}

\item{bPlot}{Boolean option determining whether create a new plot panel. \cr 
If TRUE, create a new independent plot. \cr 
If FALSE, users need to define the subplot layout first by \code{defaultplot(mfcol=c(x,y))}. See an example \code{\link{plotSimulation}} in Feisty_plotting.R.}
}
\description{
Make a plot of the adult biomass (SSB) of all functional groups
as a function of the whole simulation time.
}
\details{
X-axis is time. Y-axis is the SSB data in log10 scale.
}
\examples{
sim=simulateFeisty(simpleOutput=TRUE)
plotSSBtime(sim, bPlot=TRUE)

}
\seealso{
\code{\link{simulateFeisty}} Run FEISTY model simulations \cr
\code{\link{calcSSB}} 	Spawning stock biomass calculation
}
\author{
Yixin Zhao
}
