% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_setup.R
\name{setupBasic2}
\alias{setupBasic2}
\title{setupBasic2}
\usage{
setupBasic2(szprod = 100, 
                   lzprod = 100, 
                   bprodin = NA, 
                   dfbot = NA, 
                   depth = 100, 
                   Tp = 10, 
                   Tb = 8, 
                   nStages=9, 
                   etaMature=0.25, 
                   Fmax=0, 
                   etaF=0.05, 
                   bET=TRUE)
}
\arguments{
\item{szprod}{Small mesozooplankton production. \cr
The parameter represents small mesozooplankton carrying capacity [g/m2] and multiplied with the growth rate \bold{r}, which is always 1 [1/yr], it gives the maximum resource production [g/m2/year].
Therefore, it is described as small mesozooplankton production [g/m2/year]. \code{lzprod} and \code{bprod} are same.}

\item{lzprod}{Large mesozooplankton production.}

\item{bprodin}{Benthic production input [g/m2/year]. Default NA. Input either of `bprodin` or `dfbot`.
The benthic production `bprod` equals `bprodin`.}

\item{dfbot}{Detrital flux reaching the bottom [g/m2/year]. Default NA. Input either of `bprodin` or `dfbot`.
It will multiply the trophic transfer efficiency (10\%) to get the benthic production `bprod`. If both are NAs then `bprod = 5`.}

\item{depth}{Water column depth [meter]. depth>=200 is characterized as deep water, and depth<200 is characterized as shallow water. 
depth=300 and depth=1000 do not have any different effects on simulations.}

\item{Tp}{Pelagic water temperature, representing the top 100m average temperature [Celsius].}

\item{Tb}{Bottom water temperature [Celsius].}

\item{nStages}{Size number of large fish functional groups (e.g., large pelagic fish, demersal fish, and midwater predators). 
The size number of small fish functional groups (e.g., small pelagic fish and mesopelagic fish) is \code{round(2/3*nStages)}. 
Generally, \code{nStages} is multiples of 3 (e.g., \code{nStages = 3, 6, 9, or 12}...).}

\item{etaMature}{The coefficient determines the fish size \code{mMature} with a 50\% maturity level. 
\code{mMature = etaMature * mMax},  where \code{mMax} is the largest fish size (boundary) of a fish functional group. See \code{\link{paramAddGroup}}. 
In van Denderen et al. (2021), it was 0.002.}

\item{Fmax}{Maximum fishing mortality [1/year]. \cr
If \code{Fmax} is 0, there is no fishing mortality.\cr 
If \code{Fmax} is assigned a value greater than 0, fishing mortality will be set by multiplying the fishing selectivity \code{psi} which is based on a S-shape function. See source code of \code{\link{setFishing}}.\cr
Note here it only allows assigning fishing mortality to all functional types based on the `Fmax` input. 
If users want to assign fishing mortality to specific functional types with different `Fmax` and `etaF`, it can be done by calling \code{\link{setFishing}} later on. Details are in \code{\link{setFishing}}.}

\item{etaF}{The coefficient determining the fish size \code{mFishing} with 50\% fishing selectivity. See source code of \code{\link{setFishing}}.}

\item{bET}{Logical flag, controlling whether turns on the effective temperature effects on the large demersal fish. See Petrik et al., 2019. See `updateET` source code in FEISTY_parms.R.}
}
\value{
Additional parameters added by the function \code{\link{paramInit}}:
\itemize{
\item szprod, Small mesozooplankton production, from parameter input.
\item lzprod, Large mesozooplankton production, from parameter input.
\item bprodin, Benthic production input, from parameter input. If input is NA, the returned value is -1 for passing to FORTRAN.
\item dfbot, Detrital flux reaching the bottom, from parameter input. If input is NA, the returned value is -1 for passing to FORTRAN.
\item bprod, Benthic production, from calculation based on `bprodin` or `dfbot`.
\item depth, Water column depth, from parameter input.
\item Tp, Pelagic water temperature, from parameter input.
\item Tb, Bottom water temperature, from parameter input.
\item etaMature, The coefficient determines the fish size \code{mMature} with a 50\% maturity level, from parameter input.
\item mMedium, The boundary weight (mass) between small fish (mc <= mMedium) and medium fish (mMedium < mc < mLarge).
\item mLarge, The boundary weight (mass) between medium fish (mMedium < mc < mLarge) and large fish (mc >= mLarge).
\item bET, Logical flag, TRUE or FALSE. Effective temperature effects on the large demersal fish.
}

Added by the function \code{\link{paramSizepref}}:
\itemize{
\item theta, the size preference matrix
}

Added by the function \code{\link{setupBasic2}}:
\itemize{
\item setup, name (character) of this setup
}

Other returned parameters can be found in \code{\link{paramInit}}, \code{\link{paramAddResource}}, \code{\link{paramAddGroup}},
\code{\link{paramAddPhysiology}}, \code{\link{paramTeffect}}, \code{\link{setFishing}}, and \code{\link{paramSizepref}}.
}
\description{
\code{setupBasic2} creates a revised setup based on \code{setupBasic}.
}
\details{
The setupBasic2 makes a revised three-species setup (small pelagic fish, large pelagic fish, and demersal fish) based on Petrik et al. (2019). 
There are three resources: small mesozooplankton, large mesozooplankton, benthos, and a spare position in the state variable vector. \cr
Main revision:
\itemize{
\item Allowing more size numbers in each functional type. See \code{\link{paramAddGroup}}.
\item Generalized size-based maturity
\deqn{maturity level = (1 + ({mc}/mMature)^{-5})^{-1}}{maturity level = (1 + (mc/mMature)^(-5))^(-1)} 
where `mc` is the vector containing the geometric mean size of each class of a functional type, and `mMature` is the body size with a 50\% maturity level. \cr
See \code{\link{paramAddGroup}}.
\item Generalized size-based feeding preference
\deqn{\theta_{i,j} = \exp\left( -\frac{(\log(\frac{mc{i}}{\beta \cdot mc{_j}}))^2}{(2 \cdot \sigma)^2} \right)}{\theta[i,j] = exp(-(log(mc[i]/(beta*mc[j])))^2/(2*sigma)^2)}
\deqn{\theta_{i,j} = 0, if {mc_j} > {mc_i}}{\theta[i,j] = 0, if mc[j] > mc[i]}
where \eqn{\theta} is the size preference for a predator i to a prey j. The  'mc[i]' and 'mc[j]' are the geometric mean size of a size class of predator `i` and prey `j`.
`beta` and `sigma` are the preferred predator/prey mass ratio and the width of size preference for feeding, respectively. Default 400 and 1.3. \cr
The second equation indicates that the predator cannot prey on an organism larger than itself.
See \code{\link{paramSizepref}}.
\item Allowing the size-based fishing mortality. See \code{\link{setFishing}}.
\item Further process of size preference \eqn{\theta} for feeding preference. \cr
Fish of same size can conduct cannibalism (except medium-size demersal fish). \cr
Small pelagics and large pelagics do \bold{not} prey on benthic resources and medium-size demersal fish. \cr
The size preference for the large pelagic fish on small pelagic fish (functional type) is reduced by multiplying a coefficient 0.5 (\eqn{\theta}*0.5). \cr
Medium-size demersal fish do \bold{not} prey on zooplankton and all fish (no cannibalism as well); only eat benthic resources. \cr

Shallow water and deep water:
\itemize{
\item In shallow water (depth<200): \cr
Large demersal fish eat everything. They have reduced feeding preference on all small pelagics (\eqn{\theta}*0.5*0.75) and all large pelagics (\eqn{\theta}*0.75). \cr
Large demersal fish have feeding preference on zooplankton although the values are small. \cr

\item In deep water (depth>=200): \cr
Large-size demersal fish do not eat pelagic prey, only eat benthic resources, medium- and large-size demersals.

}

Small, medium, and large fish are differentiated according to `mMedium = 0.5g` and `mLarge = 250g`. \cr
`mMedium`: The boundary weight (mass) between small fish (mc <= mMedium) and medium fish (mMedium < mc < mLarge). \cr
`mLarge`: The boundary weight (mass) between medium fish (mMedium < mc < mLarge) and large fish (mc >= mLarge).

}
}
\examples{
p=setupBasic2(szprod = 200, 
              lzprod = 150, 
              bprodin = 15, 
              dfbot = NA, 
              depth = 300, 
              Tp = 10, 
              Tb = 9, 
              nStages=6, 
              etaMature=0.25, 
              Fmax=0, 
              etaF=0.05, 
              bET=TRUE)
sim=simulateFEISTY(p=p)
plotSimulation(sim)

}
\references{
Petrik, C. M., Stock, C. A., Andersen, K. H., van Denderen, P. D., & Watson, J. R. (2019). Bottom-up drivers of global patterns of demersal, forage, and pelagic fishes. Progress in oceanography, 176, 102124.
}
\seealso{
\code{\link{paramInit}} 	Initialize parameters for FEISTY \cr
\code{\link{paramAddResource}} 	Add resource parameters \cr
\code{\link{paramAddGroup}} 	Add parameters of one functional type \cr
\code{\link{paramAddPhysiology}} 	Add physiological parameters \cr
\code{\link{paramSizepref}} 	Size preference matrix calculation \cr
\code{\link{paramTeffect}} 	Add temperature effects \cr
\code{\link{setFishing}} 	Set fishing mortality \cr
\code{\link{simulateFEISTY}} The main function to run FEISTY simulations
}
\author{
Ken H. Andersen, Karline Soetaert, Yixin Zhao
}
