% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_setup.R
\name{setupVertical}
\alias{setupVertical}
\title{setupVertical}
\usage{
setupVertical(szprod = 80, 
                     lzprod = 80, 
                     bprodin = NA, 
                     dfbot = NA, 
                     dfpho = NA, 
                     region = 4, 
                     depth = 800, 
                     photic = 150)
}
\arguments{
\item{szprod}{Small mesozooplankton production. \cr
The parameter represents small mesozooplankton carrying capacity [g/m2] and multiplied with the growth rate \bold{r}, which is always 1 [1/yr], it gives the maximum resource production [g/m2/year].
Therefore, it is described as small mesozooplankton production [g/m2/year]. \code{lzprod} is the same.}

\item{lzprod}{Large mesozooplankton production.}

\item{bprodin}{Benthic production input [g/m2/year]. Default NA. Input either of `bprodin`, `dfbot` or `dfpho`.
The benthic production `bprod` equals `bprodin`.}

\item{dfbot}{Detrital flux reaching the bottom [g/m2/year]. Default NA. Input either of `bprodin`, `dfbot` or `dfpho`.
It will multiply the trophic transfer efficiency (10\%) to get the benthic production `bprod`.}

\item{dfpho}{Detrital flux out of the photic zone [g/m2/year]. Default NA. Input either of `bprodin`, `dfbot` or `dfpho`. If all are NAs then `dfpho = 150`.
`dfpho` will be further calculated based on the Martin curve to get detrital flux reaching the bottom and then multiplied the trophic transfer efficiency (10\%) to get benthic production `bprod` ultimately .\cr
See source code of \code{setupVertical}.}

\item{region}{Different regions: 1 Tropical, 2 Temperate, 3 Boreal, 4 Default 10 Celsius.
It represents the water column temperature profile for three regions. 
The default is 10 Celcius for the whole water column (\code{region = 4}). It is the same dataset used in van Denderen et al. (2021).}

\item{depth}{water column depth [meter]. \cr 
Different \code{depth} values will influence fish vertical overlap and temperature-dependent physiological rates. See source code of \code{setupVertical}}

\item{photic}{Photic zone depth [meter]. The value affects the diel vertical migration depth and the detrital flux reaching the bottom (if `dfpho` is used). See source code of \code{setupVertical}.}
}
\value{
Additional parameters added by function \code{\link{paramInit}}:
\itemize{
\item szprod, Small mesozooplankton production, from parameter input.
\item lzprod, Large mesozooplankton production, from parameter input.
\item bprodin, Benthic production input, from parameter input. If input is NA, the returned value is -1 for passing to FORTRAN.
\item dfbot, Detrital flux reaching the bottom, from parameter input. If input is NA, the returned value is -1 for passing to FORTRAN.
\item dfpho, Detrital flux out of the photic zone, from parameter input. If input is NA, the returned value is -1 for passing to FORTRAN.
\item bprod, Benthic production, from calculation based on `bprodin` or `dfbot`, or `dfpho`.
\item bottom, Water column depth, from parameter input (depth).
\item photic, Photic zone depth, from parameter input.
\item shelfdepth, Continental shelf depth. 250m, cannot be changed in setupVertical. 
      If the water is shallow (bottom<=shelfdepth), mesopelagic fish and midwater predators do not exist,
      and fish do not conduct diel vertical migration (dvm=0). See source code of \code{setupVertical} or \code{setupVertical2}.
\item visual, visual=1.5: visual predator. visual=1: non-visual predator.
\item etaMature, The coefficient determines the fish size \code{mMature} with a 50\% maturity level. 
It is a constant 0.002, following van Denderen et al. (2021).
\item region, Water region index, from parameter input.
}

Added by function \code{\link{paramSizepref}}:
\itemize{
\item sizepref, the size preference matrix for each predator x to each prey y.
}

Added by function \code{\link{setupVertical}}:
\itemize{
\item setup, name (character) of this setup
\item dvm, diel vertical migration depth [m]
\item ixmedium, an index indicating where medium size fish start.
\item ixlarge, an index indicating where large size fish start. E.g., ixmedium = 4, ixlarge = 7: number 1 to 3 represent small fish, number 4 to 6 represent medium fish, number 7 to the last size class represent large fish
\item depthDay, a matrix containing vertical distribution data during daytime for each resource and size class (column) in water (row)
\item dayout, a matrix containing overlap data during daytime for each predator x to each prey y
\item depthNight, a matrix containing vertical distribution data during the night for each resource and size class (column) in water (row)
\item nightout, a matrix containing overlap data during the night for each predator x to each prey y
\item vertover, the average vertical overlap matrix for each predator x to each prey y. `(dayout+nightout)/2`
\item theta, the feeding preference matrix for each predator x to each prey y. It is the product of `sizeprefer` and `vertover`.
}

Other parameters returned can be found in \code{\link{paramInit}}, \code{\link{paramAddResource}}, \code{\link{paramAddGroup}}, and
\code{\link{paramAddPhysiology}}.
}
\description{
\code{setupVertical} creates a basic five-species setup with vertical distribution as described in van Denderen et al. (2021).
}
\details{
The setupVertical makes a basic five-species setup (small pelagic fish, mesopelagic fish, large pelagic fish, midwater predators, and demersal fish) as described in van Denderen et al. (2021). 
There are three resources: small mesozooplankton, large mesozooplankton, benthos, and a spare position in the state variable vector. \cr
}
\examples{
p=setupVertical(szprod = 200, 
                lzprod = 150, 
                bprodin = NA, 
                dfbot = NA, 
                dfpho = 100, 
                region = 1, 
                depth = 1000, 
                photic = 120)
sim=simulateFEISTY(p=p)
plotSimulation(sim)

}
\references{
van Denderen, P. D., Petrik, C. M., Stock, C. A., & Andersen, K. H. (2021). Emergent global biogeography of marine fish food webs. Global Ecology and Biogeography, 30(9), 1822-1834.
}
\seealso{
\code{\link{paramInit}} 	Initialize parameters for FEISTY \cr
\code{\link{paramAddResource}} 	Add resource parameters \cr
\code{\link{paramAddGroup}} 	Add parameters of one functional type \cr
\code{\link{paramAddPhysiology}} 	Add physiological parameters \cr
\code{\link{paramSizepref}} 	Size preference matrix calculation \cr
\code{\link{simulateFEISTY}} The main function to run FEISTY simulations
}
\author{
Ken H. Andersen, Karline Soetaert, Yixin Zhao
}
