% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feisty_fishing.R
\name{calcYield}
\alias{calcYield}
\title{Yield Calculation}
\usage{
calcYield(sim, etaTime = 0.4)
}
\arguments{
\item{sim}{The FEISTY simulation result list from \code{\link{simulateFeisty}}. Note `simpleOutput` must be `TRUE` e.g., \code{sim=simulateFeisty(simpleOutput=TRUE)}.}

\item{etaTime}{The last fraction of the simulation period (default the last 40\%).}
}
\value{
Add yield data to the result list:
\itemize{
\item yieldMean: a vector containing the mean yield data [gww/m2/year] of each functional type of the time range specified.
\item yieldMin: a vector containing the minimum yield data [gww/m2/year] of each functional type within the time range specified.
\item yieldMax: a vector containing the maximum yield data [gww/m2/year] of each functional type within the time range specified.
\item yield: a matrix containing the yield data [gww/m2/year] of each functional type (column) in each time point (row)
}
}
\description{
This function calculates the yield for each function type based on a FEISTY simulation result.
}
\details{
This function calculates the yield for each function type based on the FEISTY simulation results within the specified time fraction (default is the last 40\% of the simulation period). \cr
Yield is the product of biomass \code{u} [gww/m2] and fishing mortality \code{mortF} [1/year]. Negative values in \code{u} are revised to 0. \cr
This function has been integrated into \code{\link{simulateFeisty}}. It cannot be called independently.
}
\examples{
no examples

}
\seealso{
\code{\link{setFishing}} 	Set fishing mortality
}
\author{
Yixin Zhao
}
