% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feisty_plotting.R
\name{plotDiet}
\alias{plotDiet}
\title{Diet plot}
\usage{
plotDiet(p=sim$p, u=sim$u)
}
\arguments{
\item{p}{The parameter set \code{sim$p}. The \code{simpleOutput} must be TRUE in simulateFeisty().}

\item{u}{the result matrix of the biomass, including resources and fish.\code{sim$u}.}
}
\description{
Revised Based on work of Daniel Ottmann Riera and Solenne Roux.
This function only works on four prepared setups or revised version based on these four setups.
If customized setups by users has massive differences e.g., more resources and fish functional types, this plot function may not work.
Users may need to update scripts based on source codes.
}
\examples{
sim=simulateFeisty(simpleOutput=TRUE)
plotDiet(p=sim$p, u=sim$u)

}
\seealso{
\code{\link{simulateFeisty}} Run FEISTY model simulations
}
\author{
Yixin Zhao
}
