\name{plotFEISTY}
\alias{plotSSBtime}
\alias{plotSpectra}
\alias{plotRates}
\alias{plotSimulation}
\alias{plotNetwork}
\alias{plotDiet}

\title{
Basic plotting tools for FEISTY
}
\description{
Basic plotting tools for FEISTY simulation results. Users might need to design their new plotting scripts based on these plotting tools.
}
\usage{
plotSSBtime(sim, bPlot=TRUE)
plotSpectra(sim, bPlot=TRUE)
plotRates(p=sim$p, y=p$u0, bPlot=TRUE)
plotSimulation(sim)
plotNetwork(p=sim$p, u=sim$u)
plotDiet(p=sim$p, u=sim$u)
}

\arguments{
  \item{sim}{ \bold{sim} is the data frame of simulation results, which is generated by function \code{\link{simulateFeisty}} Note the \bold{simpleOutput} option must be TRUE: \code{sim=simulateFeisty(...,simpleOutput=TRUE)}.
}

  \item{bPlot}{ if TRUE, create a new independ plot. \cr
                if FALSE, users need to define the subplot layout first by \code{defaultplot(mfcol=c(x,y))}. See an example \code{\link{plotSimulation}} in Feisty_plotting.R.
}
  \item{p}{ the parameter set \code{sim$p}.
}
  \item{y}{ the parameter set \code{sim$p}For example, \code{sim$u[,sim$nTime]}. This means the biomass data in the last time point is used for rate calculation. The time-averaged biomass data might be prefered,\code{sim$u[,sim$nTime]}.
}
  \item{u}{ the result matrix of the biomass, including resources and fish.\code{sim$u}. 
}
}
\details{
\code{\link{plotSSBtime}} plots the fish spawning stock biomass over the total simulation time.

\code{\link{plotSpectra}} plots fish biomass over size spectrum. Data is averaged data over the last 40\% simulation time.

\code{\link{plotRates}} plots growth rate (1/year), mortality (1/year), and feeding level (dimensionless) over size spectrum. Data is calculated based on averaged biomass over the last 40\% simulation time.

\code{\link{plotSimulation}} is designed for \code{\link{webFeisty}}, which includes \code{\link{plotSSBtime}}, \code{\link{plotSpectra}}, and \code{\link{plotRates}}.Additionally, there is a function \code{\link{addLegends}} in Feisty_plotting.R to add legends. However, this function is not universal for general plots. Users may need to define their new scripts for adding legends according to the function \code{\link{addLegends}}.

\code{\link{plotNetwork}} generates a food-web plot. The circle diameter represents the relative biomass level.

\code{\link{plotDiet}} shows the fraction of food sources in fish stomachs.

}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{
Yixin Zhao
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- directly executable ----

## get simulation results
sim=simulateFeisty(p  = setupBasic(), 
                    tEnd   = 200,
                    USEdll = TRUE,
                    simpleOutput = TRUE)
## plot biomass spectrum 
plotNetwork(p=sim$p, u=sim$u)
## plot spawning stock biomass over time in a new plot
plotSSBtime(sim, bPlot=TRUE)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
