% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feisty_plotting.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Food web plot}
\usage{
plotNetwork(p=sim$p, y=p$u0)
}
\arguments{
\item{p}{The parameter set \code{sim$p}. The \code{simpleOutput} must be TRUE in simulateFeisty().}

\item{u}{the result matrix of the biomass, including resources and fish.\code{sim$u}.}
}
\description{
Revised Based on the work of Daniel Ottmann Riera and Solenne Roux.
This function only works on four prepared setups or revised versions based on these four setups.
If customized setups by users have massive differences e.g., more resources and fish functional types, this plot function may not work.
Users may need to update scripts based on source codes.
}
\examples{
sim=simulateFeisty(simpleOutput=TRUE)
plotNetwork(p=sim$p, u=sim$u)

}
\seealso{
\code{\link{simulateFeisty}} Run FEISTY model simulations
}
\author{
Yixin Zhao
}
