% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_fishing.R
\name{setFishing}
\alias{setFishing}
\title{Set Fishing Mortality}
\usage{
setFishing(p, F = 0, etaF = 0.05)
}
\arguments{
\item{p}{The model parameter list, as created with e.g. \link{setupBasic2}, needed to be updated.}

\item{F}{The baseline fishing mortality rate [1/year]. Default 0, indicating no fishing.}

\item{etaF}{A coefficient determining the fish size with 50\% fishing selectivity. The value represents the fraction of the maximum size of a fish functional type. The default value is 0.05.}
}
\value{
It returns an updated parameter list:
\itemize{
\item F, baseline fishing mortality, from parameter input.
\item etaF, from parameter input.
\item mortF, a vector containing fishing mortality of all state variables, including resources (always 0) and fish.
}
}
\description{
This function sets fishing mortality for all fish based on a baseline fishing mortality and the size-based trawl selectivity.
}
\details{
The function sets fishing mortality for all fish.
For each group, it calculates the selectivity \code{psi} using the standard trawl selectivity formula from Andersen (2019) \bold{Fig 5.2}. 
The fishing mortality \code{mortF} for each group is then updated based on the calculated selectivity \code{psi} and the baseline fishing mortality rate \code{F}.
}
\examples{
p = setupBasic2(F=0) # No fishing mortality
p = setFishing(p, F = 1, etaF = 0.05) # add fishing mortality

}
\references{
Andersen, K. H. (2019). Fish ecology, evolution, and exploitation: a new theoretical synthesis. Princeton University Press.
}
\seealso{
\code{\link{calcYield}} Yield calculation
}
\author{
Yixin Zhao
}
