% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feisty_setup.R
\name{setupBasic}
\alias{setupBasic}
\title{setupBasic}
\usage{
p=setupBasic(szprod = 100, lzprod = 100, bprod = 5, depth = 100, Tp = 10, Tb = 8)
}
\arguments{
\item{szprod}{Small zooplankton productivity. \cr
Actually this represents small zooplankton carrying capacity [gww/m2] but it will multiply growth rate \bold{r} which is always 1 [1/yr]. 
Therefore, it is described as small zooplankton productivity [gww/m2/year]. \code{lzprod} and \code{bprod} are same.}

\item{lzprod}{Large zooplankton productivity.}

\item{bprod}{Benthic organism productivity.}

\item{depth}{Water column depth [meter]. depth>=200 is characterized as deep water, and depth<200 is characterized as shallow water. 
depth=300 and depth=1000 do not have any different effects on simulations.}

\item{Tp}{Pelagic water temperature, representing the top 100m average temperature [Celsius].}

\item{Tb}{Bottom water temperature [Celsius].}
}
\value{
Additional parameters added by function \code{\link{paramInit}}:
\itemize{
\item szprod, Small zooplankton productivity, from parameter input.
\item lzprod, Large zooplankton productivity, from parameter input.
\item bprod, Benthic organism productivity, from parameter input.
\item depth, Water column depth, from parameter input.
\item Tp, Pelagic water temperature, from parameter input.
\item Tb, Bottom water temperature, from parameter input.
\item mMedium, The boundary weight (mass) between small fish (mc <= mMedium) and medium fish (mMedium < mc < mLarge).
\item mLarge, The boundary weight (mass) between medium fish (mMedium < mc < mLarge) and large fish (mc >= mLarge).
}

Added by function \code{\link{setupBasic}}:
\itemize{
\item setup, name (character) of this setup
\item theta, size preference matrix added manually.
}

Other parameters returned can be found in \code{\link{paramInit}}, \code{\link{paramAddResource}},
\code{\link{paramAddGroup}}, \code{\link{paramAddPhysiology}}, and \code{\link{paramTeffect}}.
}
\description{
\code{setupBasic} creates a basic three-species setup as described in Petrik et al (2019).
}
\details{
setupBasic makes a basic three-species setup (small pelagic fish, large pelagic fish and demersal fish) as described in Petrik et al (2019). 
There are four resources including small zooplankton, large zooplankton, small benthos, and large benthos. Large benthos actually does \bold{not exist} (always 0).
}
\examples{
p=setupBasic(szprod = 200, lzprod = 150, bprod = 15, depth = 300, Tp = 10, Tb = 9)
sim=simulateFeisty(p=p, simpleOutput=TRUE)

}
\references{
Petrik, C. M., Stock, C. A., Andersen, K. H., van Denderen, P. D., & Watson, J. R. (2019). Bottom-up drivers of global patterns of demersal, forage, and pelagic fishes. Progress in oceanography, 176, 102124.
}
\seealso{
\code{\link{paramInit}} 	Initialize parameters for FEISTY \cr
\code{\link{paramAddResource}} 	Add resource parameters \cr
\code{\link{paramAddGroup}} 	Add parameters of one functional type \cr
\code{\link{paramAddPhysiology}} 	Add physiological parameters \cr
\code{\link{paramTeffect}} 	Add temperature effects \cr
\code{\link{simulateFeisty}} The main function to run FEISTY simulations
}
\author{
Ken H Andersen, Karline Soetaert <karline.soetaert@nioz.nl>, Yixin Zhao
}
