% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_setup.R
\name{setupBasic2}
\alias{setupBasic2}
\title{setupBasic2}
\usage{
p=setupBasic2(szprod = 100, lzprod = 100, bprod = 5, depth = 100, Tp = 10, Tb = 8, 
nStages=9, etaMature=0.25, F=0, etaF=0.05)
}
\arguments{
\item{szprod}{Small zooplankton productivity. \cr
Actually, this represents small zooplankton carrying capacity [gww/m2] but it will multiply the growth rate \bold{r} which is always 1 [1/yr]. 
Therefore, it is described as small zooplankton productivity [gww/m2/year]. \code{lzprod} and \code{bprod} are same.}

\item{lzprod}{Large zooplankton productivity.}

\item{bprod}{Benthic organism productivity.}

\item{depth}{Water column depth [meter]. depth>=200 is characterized as deep water, and depth<200 is characterized as shallow water. 
depth=300 and depth=1000 do not have any different effects on simulations.}

\item{Tp}{Pelagic water temperature, representing the top 100m average temperature [Celsius].}

\item{Tb}{Bottom water temperature [Celsius].}

\item{nStages}{size number of large fish functional groups (e.g., large pelagic fish, demersal fish, and bathypelagic fish). 
The size number of small fish functional groups (e.g., small  pelagic fish and mesopelagic fish) is \code{round(2/3*nStages)}. 
Generally, \code{nStages} is multiples of 3 (e.g., \code{nStages = 3, 6, 9, or 12}...).}

\item{etaMature}{The coefficient determines the fish size \code{mMature} with a 50\% maturity level. 
\code{mMature = etaMature * mMax},  where \code{mMax} is the largest fish size (boundary) of a fish functional group. See \code{\link{paramAddGroup}}. 
In van Denderen et al. (2021), it was 0.002.}

\item{F}{Baseline fishing mortality [1/year]. \cr
If \code{F} is 0, there is no fishing mortality.\cr 
If \code{F} is assigned a value greater than 0, fishing mortality will be set by multiplying the fishing selectivity \code{psi} which is based on a S-shape function.
See source code of \code{\link{setFishing}}.}

\item{etaF}{The coefficient determining the fish size \code{mFishing} with 50\% fishing selectivity. See source code of \code{\link{setFishing}}.}
}
\value{
Additional parameters added by the function \code{\link{paramInit}}:
\itemize{
\item szprod, Small zooplankton productivity, from parameter input.
\item lzprod, Large zooplankton productivity, from parameter input.
\item bprod, Benthic organism productivity, from parameter input.
\item depth, Water column depth, from parameter input.
\item Tp, Pelagic water temperature, from parameter input.
\item Tb, Bottom water temperature, from parameter input.
\item etaMature, The coefficient determines the fish size \code{mMature} with a 50\% maturity level, from parameter input.
\item mMedium, The boundary weight (mass) between small fish (mc <= mMedium) and medium fish (mMedium < mc < mLarge).
\item mLarge, The boundary weight (mass) between medium fish (mMedium < mc < mLarge) and large fish (mc >= mLarge).
}

Added by the function \code{\link{paramSizepref}}:
\itemize{
\item theta, the size preference matrix
}

Added by the function \code{\link{setupBasic2}}:
\itemize{
\item setup, name (character) of this setup
}

Other parameters returned can be found in \code{\link{paramInit}}, \code{\link{paramAddResource}}, \code{\link{paramAddGroup}},
\code{\link{paramAddPhysiology}}, \code{\link{paramTeffect}}, \code{\link{setFishing}}, and \code{\link{paramSizepref}}.
}
\description{
\code{setupBasic2} creates a revised setup based on \code{setupBasic}.
}
\details{
The setupBasic2 makes a revised three-species setup (small pelagic fish, large pelagic fish, and demersal fish) based on Petrik et al. (2019). 
There are four resources: small zooplankton, large zooplankton, small benthos, and large benthos. Large benthos actually do \bold{not exist} (always 0).\cr
Main revision:
\itemize{
\item Generalized size-based maturity
\item Generalized size-based feeding preference.
\item allowing more size numbers in each functional type.
\item allowing the size-based fishing mortality.
}
}
\examples{
p=setupBasic2(szprod = 200, lzprod = 150, bprod = 15, depth = 300, Tp = 10, Tb = 9, 
nStages=6, etaMature=0.25, F=0,etaF=0.05)
sim=simulateFEISTY(p=p)
plotSimulation(sim)

}
\references{
Petrik, C. M., Stock, C. A., Andersen, K. H., van Denderen, P. D., & Watson, J. R. (2019). Bottom-up drivers of global patterns of demersal, forage, and pelagic fishes. Progress in oceanography, 176, 102124.
}
\seealso{
\code{\link{paramInit}} 	Initialize parameters for FEISTY \cr
\code{\link{paramAddResource}} 	Add resource parameters \cr
\code{\link{paramAddGroup}} 	Add parameters of one functional type \cr
\code{\link{paramAddPhysiology}} 	Add physiological parameters \cr
\code{\link{paramSizepref}} 	Size preference matrix calculation \cr
\code{\link{paramTeffect}} 	Add temperature effects \cr
\code{\link{setFishing}} 	Set fishing mortality \cr
\code{\link{simulateFEISTY}} The main function to run FEISTY simulations
}
\author{
Ken H Andersen, Karline Soetaert <karline.soetaert@nioz.nl>, Yixin Zhao
}
