% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_setup.R
\name{setupTimeseries}
\alias{setupTimeseries}
\title{setupTimeseries}
\usage{
setupTimeseries(p = setupVertical2(),
                       tStep_ts = 1/12,
                       tSpin = NA,
                       nSpinloop = 4,
                       szbio_ts = NA,
                       lzbio_ts = NA,
                       szprod_ts = NA,
                       lzprod_ts = NA,
                       bprodin_ts = NA,
                       dfbot_ts  = NA,
                       dfpho_ts  = NA,
                       Tp_ts = NA,
                       Tm_ts = NA,
                       Tb_ts = NA,
                       Fsmp_ts = NA,
                       Fmesop_ts = NA,
                       Flgp_ts = NA,
                       Fmidwp_ts = NA,
                       Fdem_ts = NA,
                       benthosK = 80)
}
\arguments{
\item{p}{Prescribed parameter set (based on \code{setupBasic}, \code{setupBasic2}, or \code{setupVertical2}). \cr
This determines which FEISTY setup is used for time-series simulation. This defines the non-time-varying data for the simulation grid, such as water column depth (\code{depth}) and photic zone depth (\code{photic}). 
For example, if temperature is \strong{constant} over time, define it in \code{p = setupXX()} instead of providing a time-series input.}

\item{tStep_ts}{Time step for the time-series simulation [years]. \cr
This defines the resolution of the simulation. This will be used for calculating the time-series data length and creating the time sequence for solving ode. 
If input data is in monthly intervals, use \code{tStep_ts = 1/12} (1 month = 1/12 year).}

\item{tSpin}{Spin-up time span [year].\cr
This function extracts a time sequence from \code{0} to the nearest time point \eqn{\le} \code{tSpin} from the total time sequence for spinning up.}

\item{nSpinloop}{Loop number of spinning up. \cr
For instance, if \code{tSpin = 10} and \code{nSpinloop = 4}, the simulation runs 0-10 years four times to stabilize initial conditions.  
The first run starts with \code{p$u0}, while subsequent runs use the values of the last time step of the previous run as the new initial condition.}

\item{szbio_ts}{Small mesozooplankton biomass time-series [g/m2].}

\item{lzbio_ts}{Large mesozooplankton biomass time-series [g/m2].}

\item{szprod_ts}{Small mesozooplankton productivity time-series [g/m2/year].}

\item{lzprod_ts}{Large mesozooplankton productivity time-series [g/m2/year].}

\item{bprodin_ts}{Benthic productivity time-series [g/m2/year].}

\item{dfbot_ts}{time-series of detrital flux reaching the bottom [g/m2/year]. \cr
It will multiply the trophic transfer efficiency (10\%) to get benthic productivity \code{bprod_ts}.}

\item{dfpho_ts}{Detrital flux out of the photic zone time-series [g/m2/year]. Default NA. \cr
It will be further calculated based on the Martin curve to get detrital flux reaching the bottom and then multiplied the trophic transfer efficiency (10\%) to get benthic productivity \code{bprod_ts} ultimately .
See source code of \code{setupTimeseries}.\cr
Input either of \code{bprodin_ts}, \code{dfbot_ts} or \code{dfpho_ts}. If all are NAs then \code{p$bprod} is used for the simulation \code{p$r[3] = p$bprod}. In this case, remember to add benthos arguments in \code{p = setupXX()}.}

\item{Tp_ts}{Pelagic water temperature time-series, representing the top 100m average temperature [Celsius].}

\item{Tm_ts}{Mid-water temperature time-series, representing the average temperature of 500m - up to 1500m [Celsius].}

\item{Tb_ts}{Bottom water (the bottom layer) temperature time-series [Celsius].}

\item{Fsmp_ts}{Small pelagic fish maximum fishing mortality time-series [1/year].}

\item{Fmesop_ts}{Mesopelagic fish maximum fishing mortality time-series [1/year].}

\item{Flgp_ts}{Large pelagic fish maximum fishing mortality time-series [1/year].}

\item{Fmidwp_ts}{Mid-water predator maximum fishing mortality time-series [1/year].}

\item{Fdem_ts}{Demersal fish maximum fishing mortality time-series [1/year].}

\item{benthosK}{Carrying capacity of small benthos for logistic growth [g/m2]. Default is 80. 
If \code{NA}, the value remains the same as \code{p$bprod} from \code{p = setupXX()}.}
}
\value{
A modified parameter list (\code{p}) containing:\cr
\strong{Time-Series Data:}
\itemize{
 \item \code{szbio_ts}: Small mesozooplankton biomass time-series.
 \item \code{lzbio_ts}: Large mesozooplankton biomass time-series.
 \item \code{szprod_ts}: Small mesozooplankton productivity time-series.
 \item \code{lzprod_ts}: Large mesozooplankton productivity time-series.
 \item \code{bprod_ts}: Benthic productivity (derived from \code{bprodin_ts}, \code{dfbot_ts}, or \code{dfpho_ts}).
 \item \code{Tp_ts}: Pelagic water temperature time-series.
 \item \code{Tm_ts}: Mid-water temperature time-series
 \item \code{Tb_ts}: Bottom water temperature time-series.
 \item \code{Fsmp_ts}: Small pelagic fish maximum fishing mortality time-series.
 \item \code{Fmesop_ts}: Mesopelagic fish maximum fishing mortality time-series
 \item \code{Flgp_ts}: Large pelagic fish maximum fishing mortality time-series.
 \item \code{Fmidwp_ts}: Mid-water predator maximum fishing mortality time-series.
 \item \code{Fdem_ts}: Demersal fish maximum fishing mortality time-series.
 }
 
 \strong{Others:}
 \itemize{
 \item \code{tStep_ts}: Time step for the time-series simulation.
 \item \code{tEnd_ts}: Simulation period, calculated by time-series input length divided by \code{tStep_ts}
 \item \code{tSpin}: Spin-up time span
 \item \code{nSpinloop}: Loop number of spinning up.
 \item \code{bTS}: Boolean flag indicating a time-series simulation.
 \item \code{K[3]}: Overwritten by \code{benthosK}, setting the carrying capacity for small benthos.
 }
 
\strong{Note:}
The last element of each output time-series array is the replicate element of the last element in input time-series data array to ensure smooth ODE solver integration. \cr
The other elements in \code{K} are not effective since the time-series simulation has no zooplankton dynamics.
}
\description{
\code{setupTimeseries} prepares the setup for time-series simulations based on prescribed setups: 
\code{setupBasic}, \code{setupBasic2}, or \code{setupVertical2} (but not \code{setupVertical}).
}
\details{
Revisions to the FEISTY model for time-series simulations can be found in the \strong{vignette} (run \code{vignette("FEISTY")}).\cr
This function extends the prescribed setup by adding time-series data inputs and related parameters for time-series simulations.  
The main operation on each time-series data input is adding an extra element at the end, which is just a replicate of the last element of the input. It is required for time integration by the ODE solver.\cr

\itemize{
\item All time-series arrays should have the \strong{same length}.
\item \strong{Zooplankton}: Biomass values are directly available for fish consumption (no population dynamics).
\item \strong{Benthos}: Following \emph{logistic growth}, requiring productivity and carrying capacity inputs.
\item \strong{Temperature}: In R, temperature data will be used for temperature-dependent effects computation in each time step. See source code \code{derivativesFEISTYR}.
In Fortran, the temperature effects have been pre-calculated and stored in a large matrix (See source code \code{buildforcings}), which is transmitted to Fortran. The data of each time step will be called automatically.
\item \strong{Fishing mortality}: In R, fishing mortality of each size class is calculated in each time step, based on the maximum fishing mortality time-series input.
In Fortran, it is pre-calculated.
}

\strong{Important Notes:}
\itemize{
 \item Time-series data must be provided in \strong{yearly units}. 
 If the input data is in monthly units (e.g., [g/m2/month]), convert it to [g/m2/year] by multiplying by 12.
 \item Time-series inputs \strong{must not contain NAs}.
 \item \code{szbio_ts} and \code{lzbio_ts} \strong{must} be provided for fish consumption.
 \item \code{szprod_ts} and \code{lzprod_ts} \strong{must} be provided to limit fish consumption.
 }
 
\strong{Spin up:} \cr
Only one spin-up scheme is available, offering limited customization options. If a specific spin-up scheme is required, you can either:
\itemize{
\item concatenate the spin-up time series data with the simulation time series data and set \code{tSpin = NA},
\item conduct two separate runs, one for spin-up and another for simulation.
}
}
\examples{
# Two example for 1850-1859 (10 year) and 1850-2014 (165 year) of one grid. They are monthly data but converted to yearly by multiplying 12, since FEISTY units are in year.
# However in simulations, each data only run for 1/12 year to align with month.
# Photic zone depth and water column depth are not time-varying so they are in setupVertical2().
# In the example data, zooplankton biomass (Zbio) and production (Zhploss) are halved to represent small and large zooplankton.
# The information of the example data can be found in the README file in the package source file (...\FEISTY\data).

# Example of 10 years (1850-1859, including 1859) of one grid.
data(tsinput_example_1850_2014)
p = setupTimeseries(p = setupVertical2(photic = photic, depth = depth, nStages = 12),
                  tStep_ts = 1/12,
                  tSpin = NA, # no spin-up
                  nSpinloop = 4, # disabled
                  Tp_ts = Tp[1:120],
                  Tm_ts = Tm[1:120],
                  Tb_ts = Tb[1:120],
                  szbio = Zbio[1:120]/2,
                  lzbio = Zbio[1:120]/2,
                  szprod_ts = Zhploss[1:120]/2,
                  lzprod_ts = Zhploss[1:120]/2,
                  dfbot_ts  = dfbot[1:120])
# Run by R
# tStep_ts = 1/12 represents 1/12 year (month), so the data for simulation will update every 1/12 year from the time-series.
# If the user has yearly data, the argument should be tStep_ts = 1.
simR = simulateFEISTY(p = p, USEdll = F)
# Run by Fortran
simF = simulateFEISTY(p = p, USEdll = T)

plotBiomasstime(simR) # Run by R
plotBiomasstime(simF) # Run by Fortran

# Example of 165 years (1850-2014).
data(tsinput_example_1850_2014)
p = setupTimeseries(p = setupVertical2(photic = photic, depth = depth, nStages = 12),
                  tStep_ts = 1/12,
                  tSpin = 16.5, 
                  nSpinloop = 4,
                  Tp_ts = Tp,
                  Tm_ts = Tm,
                  Tb_ts = Tb,
                  szbio = Zbio/2,
                  lzbio = Zbio/2,
                  szprod_ts = Zhploss/2,
                  lzprod_ts = Zhploss/2,
                  dfbot_ts  = dfbot,
                  Fsmp_ts   = Fspel,
                  Fmesop_ts = Fmeso,
                  Flgp_ts   = Flpel,
                  Fmidwp_ts = FmidP,
                  Fdem_ts   = Fdem)
# When the simulation period is long, running by Fortran is much faster than by R.
sim = simulateFEISTY(p = p, USEdll = T) # Run by Fortran
plotBiomasstime(sim)

}
\seealso{
\code{\link{setupBasic}} The setup following Petrik et al. (2019)    \cr
\code{\link{setupBasic2}} A revised setup based on 'setupBasic'	  \cr
\code{\link{setupVertical2}} A revised setup based on 'setupVertical' \cr
\code{\link{simulateFEISTY}} The function to run FEISTY time-series simulations \cr
}
\author{
Yixin Zhao
}
