% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feisty_setup.R
\name{setupVertical}
\alias{setupVertical}
\title{setupVertical}
\usage{
p=setupVertical(szprod = 80, lzprod = 80, bent = 150, region = 4, depth = 1500, photic = 150)
}
\arguments{
\item{szprod}{Small zooplankton productivity. \cr
Actually this represents small zooplankton carrying capacity [gww/m2] but it will multiply growth rate \bold{r} which is always 1 [1/yr]. 
Therefore, it is described as small zooplankton productivity [gww/m2/year]. \code{lzprod} is same.}

\item{lzprod}{Large zooplankton productivity.}

\item{bent}{Detrital flux out of the photic zone [gww/m2/year].
\code{bent} will be further calculated based on the Martin curve to get detrital flux reaching to the bottom for driving benthic communities. See source code of \code{setupVertical}.}

\item{region}{Different regions: 1 Tropical, 2 Temperate, 3 Boreal, 4 Default 10 Celsius.
It represents the water column temperature profile for three regions. 
The default is 10 Celcius for the whole water column (\code{region = 4}). The source file is in .../data/tempdata.dat. It is the same dataset used in van Denderen et al. (2021).}

\item{depth}{water column depth [meter]. \cr 
Different \code{depth} values will influence fish vertical overlap and temperature-dependent physiological rates. See source code of \code{setupVertical}}

\item{photic}{Photic zone depth [meter]. The value affects the diel vertical migration depth. See source code of \code{setupVertical}.}
}
\value{
Additional parameters added by function \code{\link{paramInit}}:
\itemize{
\item szprod, Small zooplankton productivity, from parameter input.
\item lzprod, Large zooplankton productivity, from parameter input.
\item bent,  Detrital flux out of the photic zone, from parameter input.
\item bottom, Water column depth, from parameter input (depth).
\item photic, Photic zone depth, from parameter input.
\item mesop, mesopelagic zone depth.
\item visual, visual=1.5: visual predator. visual=1: non-visual predator.
\item etaMature, The coefficient determines the fish size \code{mMature} with a 50\% maturity level. 
It is a constant 0.002, following van Denderen et al. (2021).
\item region, water region index, from parameter input.
}

Added by function \code{\link{paramSizepref}}:
\itemize{
\item sizepref, the size preference matrix for each predator x to each prey y.
}

Added by function \code{\link{setupVertical}}:
\itemize{
\item setup, name (character) of this setup
\item dvm, diel vertical migration depth [m]
\item depthDay, a matrix containing vertical distribution data during daytime for each resource and size class (column) in water (row)
\item dayout, a matrix containing overlap data during day time for each predator x to each prey y
\item depthNight, a matrix containing vertical distribution data during night for each resource and size class (column) in water (row)
\item nightout, a matrix containing overlap data during night for each predator x to each prey y
\item vertover, the average vertical overlap matrix for each predator x to each prey y. `(dayout+nightout)/2`
\item theta, the feeding preference matrix for each predator x to each prey y. It is the product of `sizeprefer` and `vertover`.
}

Other parameters returned can be found in \code{\link{paramInit}}, \code{\link{paramAddResource}}, \code{\link{paramAddGroup}}, and
\code{\link{paramAddPhysiology}}.
}
\description{
\code{setupVertical} creates a basic five-species setup with vertical distribution as described in van Denderen et al. (2021).
}
\details{
setupVertical makes a basic five-species setup (small pelagic fish, large pelagic fish and demersal fish) as described in van Denderen et al. (2021). 
There are four resources including small zooplankton, large zooplankton, small benthos, and large benthos. Large benthos actually does \bold{not exist} (always 0).
}
\examples{
p=setupVertical(szprod = 200, lzprod = 150, bent = 100, region = 1, depth = 1000, photic = 120)
sim=simulateFeisty(p=p, simpleOutput=TRUE)

}
\references{
van Denderen, P. D., Petrik, C. M., Stock, C. A., & Andersen, K. H. (2021). Emergent global biogeography of marine fish food webs. Global Ecology and Biogeography, 30(9), 1822-1834.
}
\seealso{
\code{\link{paramInit}} 	Initialize parameters for FEISTY \cr
\code{\link{paramAddResource}} 	Add resource parameters \cr
\code{\link{paramAddGroup}} 	Add parameters of one functional type \cr
\code{\link{paramAddPhysiology}} 	Add physiological parameters \cr
\code{\link{paramSizepref}} 	Size preference matrix calculation \cr
\code{\link{simulateFeisty}} The main function to run FEISTY simulations
}
\author{
Ken H Andersen, Karline Soetaert <karline.soetaert@nioz.nl>, Yixin Zhao
}
