% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_setup.R
\name{setupVertical2}
\alias{setupVertical2}
\title{setupVertical2}
\usage{
p=setupVertical2(szprod = 80, lzprod = 80, bent = 150, nStages = 6, region = 4, depth = 1500, photic = 150, 
mesopelagic = 250, visual = 1.5, etaMature = 0.25, F = 0, etaF=0.05)
}
\arguments{
\item{szprod}{Small zooplankton productivity. \cr
Actually, this represents small zooplankton carrying capacity [gww/m2] but it will multiply the growth rate \bold{r} which is always 1 [1/yr]. 
Therefore, it is described as small zooplankton productivity [gww/m2/year]. \code{lzprod} is the same.}

\item{lzprod}{Large zooplankton productivity.}

\item{bent}{Detrital flux out of the photic zone [gww/m2/year].
\code{bent} will be further calculated based on the Martin curve to get detrital flux reaching the bottom for driving benthic communities. See source code of \code{setupVertical}.}

\item{region}{Different regions: 1 Tropical, 2 Temperate, 3 Boreal, 4 Default 10 Celsius.
It represents the water column temperature profile for three regions. 
The default is 10 Celcius for the whole water column (\code{region = 4}). The source file is in .../data/tempdata.dat. It is the same dataset used in van Denderen et al. (2021).}

\item{depth}{water column depth [meter]. \cr 
Different \code{depth} values will influence fish vertical overlap and temperature-dependent physiological rates. See source code of \code{setupVertical}}

\item{photic}{Photic zone depth [meter]. The value affects the diel vertical migration depth. See source code of \code{setupVertical}.}

\item{mesopelagic}{Mesopelagic depth [meter]. The value affects the diel vertical migration depth. See source code of \code{setupVertical} or \code{setupVertical2}.}

\item{visual}{\code{visual=1.5}: visual predator, predation ability is enhanced during the day and decreased in the twilight zone during the day. \cr 
\code{visual=1}: non-visual predator, predation abilities are equal in day and night. \cr
It must \bold{be careful} to assign other values to \code{visual}, or the setup could crash. See source code of \code{setupVertical} or \code{setupVertical2}.}

\item{etaMature}{The coefficient determines the fish size \code{mMature} with a 50\% maturity level. \code{mMature = etaMature * mMax},  where \code{mMax} is the largest fish size (boundary) of a fish functional group. See \code{\link{paramAddGroup}}. 
In van Denderen et al. (2021), it was 0.002.}

\item{F}{Baseline fishing mortality [1/year]. \cr
If \code{F} is 0, there is no fishing mortality. \cr
If \code{F} is assigned a value greater than 0, fishing mortality will be set by multiplying the fishing selectivity \code{psi} which is based on a S-shape function. See source code of \code{\link{setFishing}}.}

\item{etaF}{the coefficient determining the fish size \code{mFishing} with 50\% fishing selectivity. See source code of \code{\link{setFishing}}.}
}
\value{
Additional parameters added by function \code{\link{paramInit}}:
\itemize{
\item szprod, Small zooplankton productivity, from parameter input.
\item lzprod, Large zooplankton productivity, from parameter input.
\item bent,  Detrital flux out of the photic zone, from parameter input.
\item bottom, Water column depth, from parameter input (depth).
\item photic, Photic zone depth, from parameter input.
\item mesop, mesopelagic zone depth. from parameter input.
\item visual, visual=1.5: visual predator. visual=1: non-visual predator. from parameter input.
\item etaMature, The coefficient determines the fish size \code{mMature} with a 50\% maturity level, from parameter input.
\item region, water region index, from parameter input.
}

Added by function \code{\link{paramSizepref}}:
\itemize{
\item sizepref, the size preference matrix for each predator x to each prey y.
}

Added by function \code{\link{setupVertical}}:
\itemize{
\item setup, name (character) of this setup
\item dvm, diel vertical migration depth [m]
\item depthDay, a matrix containing vertical distribution data during daytime for each resource and size class (column) in water (row)
\item dayout, a matrix containing overlap data during daytime for each predator x to each prey y
\item depthNight, a matrix containing vertical distribution data during the night for each resource and size class (column) in water (row)
\item nightout, a matrix containing overlap data during the night for each predator x to each prey y
\item vertover, the average vertical overlap matrix for each predator x to each prey y. `(dayout+nightout)/2`
\item theta, the feeding preference matrix for each predator x to each prey y. It is the product of `sizeprefer` and `vertover`.
}

Other parameters returned can be found in \code{\link{paramInit}}, \code{\link{paramAddResource}}, \code{\link{paramAddGroup}},
\code{\link{paramAddPhysiology}}, and \code{\link{setFishing}}.
}
\description{
\code{setupVertical2} creates a revised setup based on \code{setupVertical}.
}
\details{
The setupVertical2 makes a revised five-species setup (small pelagic fish, mesopelagic fish, large pelagic fish, bathypelagic fish, and demersal fish) with vertical distribution based on van Denderen et al. (2021). 
There are four resources: small zooplankton, large zooplankton, small benthos, and large benthos. Large benthos actually do \bold{not exist} (always 0).\cr
Main revision:
\itemize{
\item Allowing more size numbers in each functional type. See \code{\link{paramAddGroup}}.
\item Generalized size-based maturity
\deqn{maturity level = (1 + ({mc}/mMature)^{-5})^{-1}}{maturity level = (1 + (mc/mMature)^(-5))^(-1)} 
where `mc` is the vector containing the geometric mean size of each class of a functional type, and `mMature` is the body size with a 50\% maturity level. \cr
See \code{\link{paramAddGroup}}.
\item Allowing the size-based fishing mortality. See \code{\link{setFishing}}.

}
}
\examples{
p=setupVertical2(szprod = 200, lzprod = 150, bent = 100, nStages = 6, region = 1, depth = 1000, photic = 120,
mesopelagic = 250, visual = 1.5, etaMature = 0.25, F = 0, etaF = 0.05)
sim=simulateFEISTY(p=p)
plotSimulation(sim)

}
\references{
van Denderen, P. D., Petrik, C. M., Stock, C. A., & Andersen, K. H. (2021). Emergent global biogeography of marine fish food webs. Global Ecology and Biogeography, 30(9), 1822-1834.
}
\seealso{
\code{\link{paramInit}} 	Initialize parameters for FEISTY \cr
\code{\link{paramAddResource}} 	Add resource parameters \cr
\code{\link{paramAddGroup}} 	Add parameters of one functional type \cr
\code{\link{paramAddPhysiology}} 	Add physiological parameters \cr
\code{\link{paramSizepref}} 	Size preference matrix calculation \cr
\code{\link{setFishing}} 	Set fishing mortality \cr
\code{\link{simulateFEISTY}} The main function to run FEISTY simulations
}
\author{
Ken H Andersen, Karline Soetaert <karline.soetaert@nioz.nl>, Yixin Zhao
}
