% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_main.R
\name{simulateFEISTY}
\alias{simulateFEISTY}
\title{Run FEISTY model simulations}
\usage{
simulateFEISTY (bCust = FALSE, p = setupBasic(), tEnd = 100, tStep  = 1, times = seq(from=0, to=tEnd, by=tStep),
       yini = p$u0, USEdll = TRUE, Rmodel = derivativesFEISTYR)
}
\arguments{
\item{bCust}{Logical flag, indicates whether to use fixed setups (FALSE) or customized setups (TRUE). \cr 
If \code{bCust} is TRUE, FEISTY simulations based on customized setups only can be done in FORTRAN, not R. \code{useDLL} input does not work.}

\item{p}{A complete parameter list. \cr 
The parameter of FEISTY setups can be one of the following: \code{\link{setupBasic}}, \code{\link{setupBasic2}}, \code{\link{setupVertical}}, and \code{\link{setupVertical2}}.
Or, modelers can customize new setups before calling \code{simulateFEISTY}.}

\item{tEnd}{The end time for the simulation [year], i.e., simulation period of FEISTY, in years.}

\item{tStep}{The time step for ODE solving output [year]. Default 1}

\item{times}{A sequence of time points for FEISTY simulations (ODEs solving), required by function \code{\link{ode}}. Generally, it needs nothing, since it will be generated by `tEnd` and `tStep` automatically. \cr
If `NA`, the function returns only the derivative (one time step running) by `Rmodel`. The `Rmodel` is \code{\link{derivativesFEISTYR}}.}

\item{yini}{A vector containing initial biomass values of all state variables (resources and all size classes). The default is imported from the setup parameter list, `p$u0`.}

\item{USEdll}{Logical flag, determining whether the ODEs are solved in FORTRAN (`TRUE`) or R (`FALSE`). \cr
The \link{deSolve} package is used for both methods.}

\item{Rmodel}{The R function for computing derivatives, defaults to \code{\link{derivativesFEISTYR}}. Generally, should not be changed, unless modelers modify the model profoundly.}
}
\value{
A list containing the simulation results:
\itemize{
\item u: a matrix of biomass of each state variable (column) at each time point (row), including resources and all size classes of functional types.
\item R: a matrix of biomass of each resource (column) at each time point (row).
\item B: a matrix of biomass of each size class (column) at each time point (row).
\item t: a vector containing all the simulation time points. From 0 to `tEnd`.
\item nTime: The number of time points.
\item USEdll: from parameter input.
\item p: the parameter list used in the simulation, the same as the input one.
\item f: A matrix containing feeding levels [-] of all size classes of functional types over each time point. Resources not included.
\item mortpred: A matrix containing a vector containing predation mortality rate [1/year] of all resources and all size classes of functional types over each time point. Resources not included.
\item g: A matrix containing the net growth rate [1/year] of all size classes of functional types over each time point. Resources not included.
\item Repro: A matrix containing the energy used for reproduction of all size classes of functional types over each time point, rate [gWW/m2/year]. Resources not included.
\item Fin: A matrix containing the biomass flux into each size class over each time point [gWW/m2/year]. Resources not included.
\item Fout: A matrix containing the biomass flux out of each size class over each time point [gWW/m2/year]. Resources not included.
\item totMort: a matrix containing the total mortality [gWW/m2/year] of each functional type over each time point, 
which includes predation mortality, background mortality, and fishing mortality.
\item totGrazing: a matrix containing the total grazing [gWW/m2/year] of each functional type over each time point, Cmax * f (maximum consumption rate * feeding level)
To be simply, the food intake before assimilation.
\item totLoss: a matrix containing all biomass loss [gWW/m2/year] of each functional type over each time point, including unassimilated food and metabolism.
They are released to environments. where is energy loss from reproduction (1-epsRepro), to be fixed.
\item totRepro: a matrix containing the total energy used for reproduction [gWW/m2] of each functional type over each time point.
\item totRecruit: a matrix containing the total recruitment [gWW/m2] of each functional type over each time point. 
TotRecruit = TotRepro * epsRepro (reproduction efficiency)
\item totBiomass: a matrix containing the total biomass [gWW/m2] of each functional type over each time point.
\item `SSBAMean`, `SSBGMean`, `SSBMin`, `SSBMax`, and `SSB` can be found in \code{\link{calcSSB}}. \cr
`yieldAMean`, `yieldGMean`, `yieldMin`, `yieldMax`, and `yield` can be found in \code{\link{calcYield}}.
}
}
\description{
\code{simulateFEISTY} runs simulations of the FEISTY model to resolve the dynamics of marine resources and fish populations over a specified time frame. \cr
It provides options for integrating ordinary differential equations in Fortran or R for four prepared setups (\code{setupBasic}, \code{setupBasic2}, \code{setupVertical}, and \code{setupVertical2}).
It also allows simulations for customized FEISTY setups.
}
\details{
The function runs the FEISTY model simulation over the specified time frame. \cr
The simulation supports published FEISTY setups and their revised versions:
\code{\link{setupBasic}}, \code{\link{setupBasic2}}, \code{\link{setupVertical}}, and \code{\link{setupVertical2}}, and customized setups by modelers. \cr
The simulation can be conducted by either a FORTRAN-based approach or an R-based approach. Both methods rely on the \link{deSolve} package for ODE solving. 
For efficiency, FORTRAN dll should be used. For model development, the R-version is preferred.
Simulations based on customized setups only can be done by the FORTRAN-based approach.
}
\examples{
# Just some examples, data input and output may not make sense.


#-----------------------------------------------
# run setupBasic with default parameter settings
#-----------------------------------------------
sim <- simulateFEISTY().
plotSimulation(sim)

# -------------------------------------------------------------------------------

# run FEISTY simulation based on setupVertical2
# prepare a parameter list
p_V <- setupVertical2(szprod = 100, lzprod = 120, bprodin = 200, depth = 1000)
# run the simulation by R
sim_Vertical_R <- simulateFEISTY(p = p_V)
plotSimulation(sim_Vertical_R)

# run FEISTY simulation based on setupBasic2 by Fortran
sim_Basic2_F <- simulateFEISTY(p = setupBasic2(szprod = 90, lzprod = 100, bprod = 15, depth = 500, Tp = 11, Tb = 9, 
nStages=9, etaMature=0.25, F=0, etaF=0.05), tEnd = 1000, tStep = 1, USEdll = TRUE)

plotSimulation(sim_Basic2_F)

# -------------------------------------------------------------------------------

# run FEISTY simulation based on a customized set up

# Initialize the parameter list.
p_cust <- paramInit()

# add three resources
p_cust <- paramAddResource(p_cust,
          names= c("smallZoo", "largeZoo", "benthos"),
          K    = c(100, 120, 80),
          r    = c(1, 1, 1),
          mLower = c(2e-06,0.001, 0.5e-03),
          mUpper = c(0.001, 0.5, 125),
          mc   = c(2e-06*sqrt(500), 0.001*sqrt(500), 0.5e-03*sqrt(250000)))
          
# add two functional types of fish: small pelagic fish and demersal fish           
p_cust <- paramAddGroup(p_cust, mMin=0.001, mMax=250, mMature=NA, 
                        mortF=0,      nStages=6, name="smallPel")
p_cust <- paramAddGroup(p_cust, mMin=0.001, mMax=125000, mMature=NA, 
                        mortF=0, nStages=9, name="demersals")
                        
# add physiological parameters for two functional types
p_cust <- paramAddPhysiology(p_cust, 
          ac = 20, bc = -0.25,       
          am = 0.011*365, bm = -0.175,      
          ae = 70, be = -0.2,        
          epsRepro = 0.01, 
          epsAssim = 0.7)

# Add fishing mortality on demersal fish only. The baseline fishing mortality is 0.2/year.
p_cust <- setFishing(p_cust, F=0.2, etaF=0.05,groupidx=c(2))

# Add size preference
p_cust$theta <- paramSizepref(p = p_cust,   
                        beta = 400,
                        sigma = 1.3, 
                        type = 1)

# Add temperature effect
p_cust=paramTeffect(p=p_cust, Tref=10, Q10=1.88, Q10m=2.35, pelgroupidx=c(1), demgroupidx=c(2))
# Turn on the effective temperature effects on large demersals.
p_cust$bET=TRUE

# run the simulation for 500 years. 
sim_cust <- simulateFEISTY(bCust = TRUE, p = p_cust, tEnd = 500)

plotSimulation(sim_cust)

}
\references{
Petrik, C. M., Stock, C. A., Andersen, K. H., van Denderen, P. D., & Watson, J. R. (2019). Bottom-up drivers of global patterns of demersal, forage, and pelagic fishes. Progress in oceanography, 176, 102124.

van Denderen, P. D., Petrik, C. M., Stock, C. A., & Andersen, K. H. (2021). Emergent global biogeography of marine fish food webs. Global Ecology and Biogeography, 30(9), 1822-1834.

de Roos, A. M., Schellekens, T., Van Kooten, T., Van De Wolfshaar, K., Claessen, D., & Persson, L. (2008). Simplifying a physiologically structured population model to a stage-structured biomass model. Theoretical population biology, 73(1), 47-62.

Soetaert, K., Petzoldt, T., & Setzer, R. W. (2010). Solving differential equations in R: package deSolve. Journal of statistical software, 33, 1-25.
}
\seealso{
\code{\link{setupBasic}} The setup following Petrik et al. (2019) \cr
\code{\link{setupBasic2}} A revised setup based on `setupBasic` \cr
\code{\link{setupVertical}} The setup following van Denderen et al. (2021) \cr
\code{\link{setupVertical2}} A revised setup based on `setupVertical` \cr

\code{\link{calcSSB}} Spawning stock biomass calculation \cr
\code{\link{calcYield}} Yield calculation

\code{\link{derivativesFEISTYR}} The derivative function of state variables in FEISTY model

\code{\link{webFEISTY}} A shiny interface for visualizing FEISTY model results

\code{\link{plotSimulation}} Plot simulation results including rates, biomass, and SSB data
}
\author{
Ken H. Andersen, Karline Soetaert <karline.soetaert@nioz.nl>, Yixin Zhao
}
