% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webFEISTY.R
\name{webFEISTY}
\alias{webFEISTY}
\title{Visualization interface for the FishErIes Size and functional TYpe model.}
\usage{
webFEISTY(...)
}
\arguments{
\item{...}{No arguments are required for this shiny application. See Examples below.}
}
\value{
This Shiny application does not return values.
}
\description{
Launch a Shiny application, which provides a user-friendly interactive and visualization interface for the FEISTY model. Four FEISTY setups can be simulated.
There are three tabs displaying basic results, food web plots, and diet plots. Please see Details below.
}
\details{
Four FEISTY setups are \code{\link{setupBasic}}, \code{\link{setupBasic2}}, \code{\link{setupVertical}}, \code{\link{setupVertical2}} in FEISTY_setup.R.
Users can customize their FEISTY setups by changing the widgets on the left part of the Shiny interface and get the visualized results.

Tab 'Rates' is generated by \code{\link{plotSimulation}} in FEISTY_plotting.R, which mainly includes \code{\link{plotSSBtime}}, \code{\link{plotSpectra}}, and \code{\link{plotRates}}.

Tab 'Network' shows food web results based on the averaged data of the last 40\% of simulation time. See \code{\link{plotNetwork}} in FEISTY_plotting.R.

Tab 'Diet' shows fish diet results based on the averaged data of the last 40\% of simulation time. See \code{\link{plotDiet}} in FEISTY_plotting.R.
}
\examples{
webFEISTY()

}
\seealso{
\code{\link{setupBasic}} The setup following Petrik et al. (2019) \cr
\code{\link{setupBasic2}} A revised setup based on `setupBasic` \cr
\code{\link{setupVertical}} The setup following van Denderen et al. (2021) \cr
\code{\link{setupVertical2}} A revised setup based on `setupVertical`

\code{\link{simulateFEISTY}} The main function to run FEISTY simulations \cr

\code{\link{plotSimulation}} Plot simulation results \cr
\code{\link{plotSSBtime}} Spawning stock biomass plot \cr
\code{\link{plotSpectra}} Biomass plot \cr
\code{\link{plotRates}} Plots for growth rate, mortality, and feeding level \cr
\code{\link{plotNetwork}} Food web plot \cr
\code{\link{plotDiet}} Diet plot

...
}
\author{
Karline Soetaert, Ken H Andersen, Yixin Zhao
}
